location of project delete button
