+++ app/src/features/projects/view/ProjectPage.tsx
@@ -9,7 +9,7 @@ import Paper from '@material-ui/core/Paper';
 import { Theme } from '@material-ui/core/styles/createMuiTheme';
 import Typography from '@material-ui/core/Typography';
 import makeStyles from '@material-ui/core/styles/makeStyles';
-import { mdiClipboardCheckMultipleOutline, mdiInformationOutline, mdiPaperclip } from '@mdi/js';
+import { mdiClipboardCheckMultipleOutline, mdiTrashCanOutline, mdiInformationOutline, mdiPaperclip } from '@mdi/js';
 import Icon from '@mdi/react';
 import { DATE_FORMAT } from 'constants/dateTimeFormats';
 import ProjectAttachments from 'features/projects/view/ProjectAttachments';
@@ -18,10 +18,13 @@ import SurveysListPage from 'features/surveys/list/SurveysListPage';
 import { useBiohubApi } from 'hooks/useBioHubApi';
 import { IGetAllCodeSetsResponse } from 'interfaces/useCodesApi.interface';
 import { IGetProjectForViewResponse } from 'interfaces/useProjectApi.interface';
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useCallback, useEffect, useState, useContext } from 'react';
 import { useLocation, useParams } from 'react-router';
 import { NavLink } from 'react-router-dom';
 import { getFormattedDateRangeString } from 'utils/Utils';
+import Button from '@material-ui/core/Button';
+import { DialogContext } from 'contexts/dialogContext';
+import { useHistory } from 'react-router';
 
 const useStyles = makeStyles((theme: Theme) => ({
   projectNav: {
@@ -50,8 +53,10 @@ const useStyles = makeStyles((theme: Theme) => ({
 const ProjectPage: React.FC = () => {
   const urlParams = useParams();
   const location = useLocation();
-
   const biohubApi = useBiohubApi();
+  const history = useHistory();
+
+  const dialogContext = useContext(DialogContext);
 
   const classes = useStyles();
 
@@ -97,6 +102,44 @@ const ProjectPage: React.FC = () => {
     }
   }, [isLoadingProject, projectWithDetails, getProject]);
 
+  const defaultYesNoDialogProps = {
+    dialogTitle: 'Delete Project',
+    dialogText: 'Are you sure you want to delete this project, its attachments and associated surveys/observations?',
+    open: false,
+    onClose: () => dialogContext.setYesNoDialog({ open: false }),
+    onNo: () => dialogContext.setYesNoDialog({ open: false }),
+    onYes: () => dialogContext.setYesNoDialog({ open: false })
+  };
+
+  const showDeleteProjectDialog = () => {
+    dialogContext.setYesNoDialog({
+      ...defaultYesNoDialogProps,
+      open: true,
+      onYes: () => {
+        deleteProject();
+        dialogContext.setYesNoDialog({ open: false });
+      }
+    });
+  };
+
+  const deleteProject = async () => {
+    if (!projectWithDetails) {
+      return;
+    }
+
+    try {
+      const response = await biohubApi.project.deleteProject(projectWithDetails.id);
+
+      if (!response) {
+        return;
+      }
+
+      history.push(`/projects`);
+    } catch (error) {
+      return error;
+    }
+  };
+
   if (!codes || !projectWithDetails) {
     return <CircularProgress className="pageProgress" size={40} />;
   }
@@ -105,32 +148,44 @@ const ProjectPage: React.FC = () => {
     <>
       <Paper elevation={2} square={true}>
         <Container maxWidth="xl">
-          <Box py={4}>
-            <Box mb={1}>
-              <Typography variant="h1">{projectWithDetails.project.project_name}</Typography>
+          <Box display="flex" justifyContent="space-between">
+            <Box py={4}>
+              <Box mb={1}>
+                <Typography variant="h1">{projectWithDetails.project.project_name}</Typography>
+              </Box>
+              <Box>
+                <Typography variant="subtitle1" color="textSecondary">
+                  <span>
+                    {projectWithDetails.project.end_date ? (
+                      <>
+                        {getFormattedDateRangeString(
+                          DATE_FORMAT.ShortMediumDateFormat,
+                          projectWithDetails.project.start_date,
+                          projectWithDetails.project.end_date
+                        )}
+                      </>
+                    ) : (
+                      <>
+                        <span>Start Date:</span>{' '}
+                        {getFormattedDateRangeString(
+                          DATE_FORMAT.ShortMediumDateFormat,
+                          projectWithDetails.project.start_date
+                        )}
+                      </>
+                    )}
+                  </span>
+                </Typography>
+              </Box>
             </Box>
-            <Box>
-              <Typography variant="subtitle1" color="textSecondary">
-                <span>
-                  {projectWithDetails.project.end_date ? (
-                    <>
-                      {getFormattedDateRangeString(
-                        DATE_FORMAT.ShortMediumDateFormat,
-                        projectWithDetails.project.start_date,
-                        projectWithDetails.project.end_date
-                      )}
-                    </>
-                  ) : (
-                    <>
-                      <span>Start Date:</span>{' '}
-                      {getFormattedDateRangeString(
-                        DATE_FORMAT.ShortMediumDateFormat,
-                        projectWithDetails.project.start_date
-                      )}
-                    </>
-                  )}
-                </span>
-              </Typography>
+            <Box ml={4} mt={4} mb={4}>
+              <Button
+                variant="outlined"
+                color="primary"
+                data-testid="delete-project-button"
+                startIcon={<Icon path={mdiTrashCanOutline} size={1} />}
+                onClick={showDeleteProjectDialog}>
+                Delete Project
+              </Button>
             </Box>
           </Box>
         </Container>
